Remove unnecessary app wrapper after the refactor.
